# backend/docs/openapi.yaml
openapi: 3.1.0
info:
  title: Shiluv API
  version: 0.1.0
  description: >
    Suggested API contract for the weekly shift schedule and duty roster
    workflow. The initial FastAPI backend uses development-only HTTP Basic
    authentication while the frontend and roster generation logic are being
    developed, but the backend remains the authority for roles, deadlines,
    validation, and authorization.
servers:
  - url: http://localhost:8000
    description: Local FastAPI development server
  - url: https://api.example.invalid
    description: Production placeholder
...
security:
  - BasicAuth: []
paths:
...
  /api/v1/session:
    get:
      tags:
        - Users
      operationId: getSession
      summary: Get the authenticated development user
      responses:
        "200":
          description: Authenticated user session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Session"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/users:
    get:
      tags:
        - Users
      operationId: listUsers
      summary: List workers and managers
      parameters:
        - name: role
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/UserRole"
      responses:
        "200":
          description: Users
          content:
            application/json:
              schema:
                type: object
                required:
                  - users
                properties:
                  users:
                    type: array
                    items:
                      $ref: "#/components/schemas/User"
  /api/v1/weeks:
    get:
      tags:
        - Weeks
      operationId: listWeeks
      summary: List known weekly cycles
      parameters:
        - name: from
          in: query
          required: false
          description: Earliest week start date to include.
          schema:
            type: string
            format: date
        - name: to
          in: query
          required: false
          description: Latest week start date to include.
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Weekly cycles
          content:
            application/json:
              schema:
                type: object
                required:
                  - weeks
                properties:
                  weeks:
                    type: array
                    items:
                      $ref: "#/components/schemas/Week"
  /api/v1/weeks/{weekStart}:
    get:
      tags:
        - Weeks
      operationId: getWeek
      summary: Get one weekly cycle
      parameters:
        - $ref: "#/components/parameters/WeekStart"
      responses:
        "200":
          description: Weekly cycle
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Week"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/weeks/{weekStart}/programatsia:
    get:
      tags:
        - Programatsia
      operationId: getProgramatsiaStatus
      summary: Get upload and validation status for the programatsia file
      parameters:
        - $ref: "#/components/parameters/WeekStart"
      responses:
        "200":
          description: Programatsia status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProgramatsiaStatus"
        "404":
          $ref: "#/components/responses/NotFound"

    post:
      tags:
        - Programatsia
      operationId: uploadProgramatsia
      summary: Upload the weekly programatsia .xlsx file
      description: >
        Development endpoint for manager uploads. The server accepts a non-empty
        `.xlsx` file, records the upload status, and synchronously converts it
        into an available shift schedule using the student-exercise business
        logic scaffold. Production code must replace the scaffold with real
        workbook parsing and validation.
      parameters:
        - $ref: "#/components/parameters/WeekStart"
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
              properties:
                file:
                  type: string
                  format: binary
                  description: The weekly programatsia .xlsx file.
      responses:
        "202":
          description: Upload accepted and converted by the development scaffold
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProgramatsiaStatus"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "422":
          $ref: "#/components/responses/ValidationFailed"
  /api/v1/weeks/{weekStart}/schedule:
    get:
      tags:
        - Schedule
      operationId: getShiftSchedule
      summary: Get the converted shift schedule
      parameters:
        - $ref: "#/components/parameters/WeekStart"
      responses:
        "200":
          description: Shift schedule
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ShiftSchedule"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /api/v1/weeks/{weekStart}/shift-exclusions:
    get:
      tags:
        - Exclusions
      operationId: listShiftExclusions
      summary: List shift exclusions for the week
      parameters:
        - $ref: "#/components/parameters/WeekStart"
        - name: worker_id
          in: query
          required: false
          description: Limit results to one worker. Workers may only query themselves.
          schema:
            type: string
      responses:
        "200":
          description: Shift exclusions
          content:
            application/json:
              schema:
                type: object
                required:
                  - exclusions
                properties:
                  exclusions:
                    type: array
                    items:
                      $ref: "#/components/schemas/ShiftExclusion"
        "403":
          $ref: "#/components/responses/Forbidden"

    post:
      tags:
        - Exclusions
      operationId: createShiftExclusion
      summary: Exclude a worker from a shift before the cutoff
      parameters:
        - $ref: "#/components/parameters/WeekStart"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ShiftExclusionCreate"
      responses:
        "201":
          description: Shift exclusion created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ShiftExclusion"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
  /api/v1/weeks/{weekStart}/shift-exclusions/{exclusionId}:
    delete:
      tags:
        - Exclusions
      operationId: deleteShiftExclusion
      summary: Remove a shift exclusion before the cutoff
      parameters:
        - $ref: "#/components/parameters/WeekStart"
        - name: exclusionId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Shift exclusion removed
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /api/v1/weeks/{weekStart}/roster:
    get:
      tags:
        - Roster
      operationId: getDutyRoster
      summary: Get the generated duty roster
      parameters:
        - $ref: "#/components/parameters/WeekStart"
      responses:
        "200":
          description: Duty roster
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DutyRoster"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /api/v1/weeks/{weekStart}/roster/generate:
    post:
      tags:
        - Roster
      operationId: generateDutyRoster
      summary: Generate or regenerate the duty roster
      description: >
        The scheduled roster generation job should normally call this endpoint
        or the same underlying service function. Manager-triggered regeneration
        is useful during development and review.
      parameters:
        - $ref: "#/components/parameters/WeekStart"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateRosterRequest"
      responses:
        "202":
          description: Roster generation started or completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DutyRoster"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
  /api/v1/weeks/{weekStart}/roster/authorize:
    post:
      tags:
        - Roster
      operationId: authorizeDutyRoster
      summary: Authorize the duty roster after manager review
      parameters:
        - $ref: "#/components/parameters/WeekStart"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthorizeRosterRequest"
      responses:
        "200":
          description: Roster authorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DutyRoster"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /api/v1/weeks/{weekStart}/alerts:
    get:
      tags:
        - Alerts
      operationId: listAlerts
      summary: List operational alerts for a week
      parameters:
        - $ref: "#/components/parameters/WeekStart"
        - name: status
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/AlertStatus"
      responses:
        "200":
          description: Alerts
          content:
            application/json:
              schema:
                type: object
                required:
                  - alerts
                properties:
                  alerts:
                    type: array
                    items:
                      $ref: "#/components/schemas/Alert"
        "403":
          $ref: "#/components/responses/Forbidden"
...
components:
  securitySchemes:
    BasicAuth:
      type: http
      scheme: basic
      description: Development-only HTTP Basic authentication.
...
  schemas:
...
    AlertStatus:
      type: string
      enum:
        - open
        - acknowledged
        - resolved
...
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
              additionalProperties: true
...
    Session:
      type: object
      required:
        - user
      properties:
        user:
          $ref: "#/components/schemas/User"
    User:
      type: object
      required:
        - id
        - username
        - display_name
        - role
        - active
      properties:
        id:
          type: string
        username:
          type: string
        display_name:
          type: string
        role:
          $ref: "#/components/schemas/UserRole"
        active:
          type: boolean
    UserRole:
      type: string
      enum:
        - worker
        - manager
...